move mesh generation code to mesh.c
