+++ include/render.h
@@ -8,6 +8,5 @@ void render_cube();
 void r_init(shader_program* program);
 void r_cleanup();
 void render(camera cam, shader_program program);
-void update_chunk_packet_at(int x, int z);
 
 #endif
