+++ include/settings.h
@@ -5,10 +5,12 @@
 #define CHUNK_SIZE 16
 #define CHUNK_HEIGHT 64
 
-#define CHUNK_RENDER_DISTANCE 16
+#define CHUNK_RENDER_DISTANCE 12
 #define RENDER_DISTANCE 600.0f
 
 // #define WIREFRAME
+
+#define USE_MIPMAP
 
 #define SEED 123456789
