+++ src/block.c
@@ -1,17 +1,9 @@
 #include <block.h>
-#include <render.h>
+// #include <render.h>
+#include <mesh.h>
 #include <world.h>
 #include <cglm/cglm.h>
 
-void update_chunks(int x, int z) {
-    // update chunk and adjacent ones
-    update_chunk_packet_at(x, z);
-    update_chunk_packet_at(x + 1, z);
-    update_chunk_packet_at(x - 1, z);
-    update_chunk_packet_at(x, z + 1);
-    update_chunk_packet_at(x, z - 1);
-
-}
 
 float get_empty_dist(camera cam) {
     vec3 position = {cam.position[0], cam.position[1], cam.position[2]};
@@ -64,7 +56,7 @@ void break_block(camera cam) {
 
     c->blocks[chunk_x][chunk_y][chunk_z] = NULL;
     
-    update_chunks(c->x, c->z);
+    update_chunk_mesh(c->x, c->z);
 }
 
 void place_block(camera cam) {
@@ -92,10 +84,10 @@ void place_block(camera cam) {
         return;
     }
 
-    c->blocks[chunk_x][chunk_y][chunk_z] = &TYPES[6];
+    c->blocks[chunk_x][chunk_y][chunk_z] = &TYPES[8];
 
     // update chunk and adjacent chunks
-    update_chunks(c->x, c->z);
+    update_chunk_mesh(c->x, c->z);
 }
 
 
@@ -203,6 +195,19 @@ block_type TYPES[] = {
             {10.0f, 0.0f},
             {10.0f, 0.0f},
             {10.0f, 0.0f},
+        }
+    },
+    {
+        .id = 8,
+        .name = "glass",
+        .transparent = 1,
+        .face_atlas_coords = {
+            {11.0f, 0.0f},
+            {11.0f, 0.0f},
+            {11.0f, 0.0f},
+            {11.0f, 0.0f},
+            {11.0f, 0.0f},
+            {11.0f, 0.0f},
         }
     }
 };
