+++ src/render.c
@@ -8,29 +8,14 @@
 #include <assert.h>
 #include <hashmap.h>
 
+#include <mesh.h>
+
 #include <settings.h>
 
-#define VBO_WIDTH 7
-#define VERTS_PER_SIDE 6
-#define SIDE_OFFSET VERTS_PER_SIDE * VBO_WIDTH
-
-#define INITIAL_VBO_SIZE 5000 * SIDE_OFFSET
 
 VAO vao;
 VBO vbo;
 shader_program program;
-
-typedef struct {
-    float* opaque_side_data;
-    float* transparent_side_data;
-    int num_opaque_sides;
-    int num_transparent_sides;
-} chunk_packet;
-
-DEFINE_HASHMAP(chunk_packet_map, chunk_coord, chunk_packet, chunk_hash, chunk_equals);
-typedef chunk_packet_map_hashmap chunk_packet_map;
-
-chunk_packet_map packets;
 
 void r_init(shader_program* program) {
     glEnable(GL_DEPTH_TEST);
@@ -62,8 +47,7 @@ void r_init(shader_program* program) {
     // world
     w_init();
 
-    // initialize chunk packet map
-    packets = chunk_packet_map_init(CHUNK_CACHE_SIZE);
+    m_init();
 
     glViewport(0, 0, WIDTH, HEIGHT);
 }
@@ -75,100 +59,6 @@ void r_cleanup() {
     t_cleanup();
 }
 
-void pack_side(int x_0, int y_0, int z_0, uint side, block_type* type, float* side_data) {
-    int cube_vertices_offset = side * VERTS_PER_SIDE * CUBE_VERTICES_WIDTH;
-    for (int i = 0; i < 6; i++) {
-        int index = cube_vertices_offset + i * CUBE_VERTICES_WIDTH;
-        float x = x_0 + CUBE_VERTICES[index + 0];
-        float y = y_0 + CUBE_VERTICES[index + 1];
-        float z = z_0 + CUBE_VERTICES[index + 2];
-
-        float tx = CUBE_VERTICES[index + 3];
-        float ty = CUBE_VERTICES[index + 4];
-
-        int opaque_side_data_offset = i * VBO_WIDTH;
-        side_data[opaque_side_data_offset + 0] = x;
-        side_data[opaque_side_data_offset + 0] = x;
-        side_data[opaque_side_data_offset + 0] = x;
-        side_data[opaque_side_data_offset + 0] = x;
-        side_data[opaque_side_data_offset + 0] = x;
-        side_data[opaque_side_data_offset + 1] = y;
-        side_data[opaque_side_data_offset + 2] = z;
-        side_data[opaque_side_data_offset + 3] = tx;
-        side_data[opaque_side_data_offset + 4] = ty;
-        side_data[opaque_side_data_offset + 5] = type->face_atlas_coords[side][0] / (float)TEXTURE_ATLAS_SIZE;
-        side_data[opaque_side_data_offset + 6] = type->face_atlas_coords[side][1] / (float)TEXTURE_ATLAS_SIZE;
-    }
-}
-
-void pack_block(
-    int x, int y, int z,
-    chunk* c,
-    chunk* adj_chunks[4], // front, back, left, right
-    float** side_data, 
-    int* num_sides) {
-    
-    int world_x = x + (CHUNK_SIZE * c->x);
-    int world_y = y;
-    int world_z = z + (CHUNK_SIZE * c->z);
-
-    for (int side = 0; side < 6; side++) {
-        chunk* adj = NULL;
-        if (side < 4) {
-            adj = adj_chunks[side];
-        }
-
-        if (!get_side_visible(x, y, z, side, c, adj)) {
-            continue;
-        }
-
-        int new_side_count = (*num_sides) + 1;
-
-        // check if we need to reallocate memory
-        if (new_side_count * SIDE_OFFSET > INITIAL_VBO_SIZE) {
-            float* tmp = realloc(*side_data, new_side_count * SIDE_OFFSET * sizeof(float));
-            assert(tmp != NULL && "Failed to allocate memory for side data");
-            *side_data = tmp;
-        }
-        block_type* type = c->blocks[x][y][z];
-
-        pack_side(
-            world_x, world_y, world_z, 
-            side, 
-            type,
-            (*side_data) + (*num_sides) * SIDE_OFFSET
-        );
-        (*num_sides)++;
-    }
-}
-
-void pack_chunk(chunk* c, chunk* adj_chunks[4], 
-    float** opaque_side_data, int* num_opaque_sides,
-    float** transparent_side_data, int* num_transparent_sides) {
-    if (c == NULL) {
-        return;
-    }
-
-    for (int i = 0; i < CHUNK_SIZE; i++) {
-        for (int j = 0; j < CHUNK_SIZE; j++) {
-            for (int k = 0; k < CHUNK_HEIGHT; k++) {
-                if (c->blocks[i][k][j] == NULL) {
-                    continue;
-                }
-
-                if (c->blocks[i][k][j]->transparent) {
-                    pack_block(i, k, j, c, adj_chunks, 
-                        transparent_side_data, num_transparent_sides);
-                }
-                else {
-                    pack_block(i, k, j, c, adj_chunks, 
-                        opaque_side_data, num_opaque_sides);
-                }
-            }
-        }
-    }
-}
-
 void render_sides(float* side_data, int num_sides) {
     bind_vao(vao);
     buffer_data(vbo, GL_STATIC_DRAW, side_data, num_sides * SIDE_OFFSET * sizeof(float));
@@ -178,89 +68,6 @@ void render_sides(float* side_data, int num_sides) {
     use_vbo(vbo);
 
     glDrawArrays(GL_TRIANGLES, 0, num_sides * VERTS_PER_SIDE);
-}
-
-chunk_packet* create_chunk_packet(int x, int z) {
-    chunk_packet* packet = malloc(sizeof(chunk_packet));
-    assert(packet != NULL && "Failed to allocate memory for packet");
-
-    // get chunk and adjacent chunks
-    chunk* c = get_chunk(x, z);
-    chunk* adj_chunks[4] = {
-        get_chunk(x + 1, z),
-        get_chunk(x - 1, z),
-        get_chunk(x, z - 1),
-        get_chunk(x, z + 1)
-    };
-
-    // pack chunk data into packet
-    int packet_transparent_side_count = 0;
-    int packet_opaque_side_count = 0;
-
-    float* packet_opaque_sides = malloc(INITIAL_VBO_SIZE * sizeof(float));
-    float* packet_transparent_sides = malloc(INITIAL_VBO_SIZE * sizeof(float));
-    assert(packet_opaque_sides != NULL && "Failed to allocate memory for opaque packet sides");
-    assert(packet_transparent_sides != NULL && "Failed to allocate memory for transparent packet sides");
-
-    pack_chunk(c, adj_chunks, 
-        &packet_opaque_sides, &packet_opaque_side_count,
-        &packet_transparent_sides, &packet_transparent_side_count);
-
-    assert(packet != NULL && "Failed to allocate memory for packet");
-    packet->opaque_side_data = packet_opaque_sides;
-    packet->num_opaque_sides = packet_opaque_side_count;
-    packet->transparent_side_data = packet_transparent_sides;
-    packet->num_transparent_sides = packet_transparent_side_count;
-
-    chunk_coord coord = {x, z};
-    chunk_packet_map_insert(&packets, coord, *packet);
-    return packet;
-}
-
-chunk_packet* get_chunk_packet(int x, int z) {
-    chunk_coord coord = {x, z};
-    chunk_packet* packet = chunk_packet_map_get(&packets, coord);
-    if (packet == NULL) {
-        packet = create_chunk_packet(x, z);
-    }
-    return packet;
-}
-
-void update_chunk_packet_at(int x, int z) {
-    chunk_coord coord = {x, z};
-    chunk_packet* packet = chunk_packet_map_get(&packets, coord);
-    if (packet == NULL) {
-        assert(false && "Packet does not exist");
-    }
-
-    free(packet->opaque_side_data);
-    chunk_packet_map_remove(&packets, coord);
-
-    chunk* c = get_chunk(x, z);
-    chunk* adj_chunks[4] = {
-        get_chunk(x + 1, z),
-        get_chunk(x - 1, z),
-        get_chunk(x, z - 1),
-        get_chunk(x, z + 1)
-    };
-    int transparent_side_count = 0;
-    int opaque_side_count = 0;
-    float* transparent_sides = malloc(INITIAL_VBO_SIZE * sizeof(float));
-    float* opaque_sides = malloc(INITIAL_VBO_SIZE * sizeof(float));
-    assert(transparent_sides != NULL && "Failed to allocate memory for transparent packet sides");
-    assert(opaque_sides != NULL && "Failed to allocate memory for opaque packet sides");
-    pack_chunk(c, adj_chunks, 
-        &opaque_sides, &opaque_side_count, 
-        &transparent_sides, &transparent_side_count);
-
-    packet = malloc(sizeof(chunk_packet));
-    assert(packet != NULL && "Failed to allocate memory for packet");
-
-    packet->opaque_side_data = opaque_sides;
-    packet->num_opaque_sides = opaque_side_count;
-    packet->transparent_side_data = transparent_sides;
-    packet->num_transparent_sides = transparent_side_count;
-    chunk_packet_map_insert(&packets, coord, *packet);
 }
 
 void render(camera cam, shader_program program) {
@@ -284,13 +91,13 @@ void render(camera cam, shader_program program) {
     int player_chunk_x = (int)(cam.position[0] / CHUNK_SIZE);
     int player_chunk_z = (int)(cam.position[2] / CHUNK_SIZE);
 
-    chunk_packet* packets[2 * CHUNK_RENDER_DISTANCE][2 * CHUNK_RENDER_DISTANCE];
+    chunk_mesh* packets[2 * CHUNK_RENDER_DISTANCE][2 * CHUNK_RENDER_DISTANCE];
 
     for (int i = 0; i < 2 * CHUNK_RENDER_DISTANCE; i++) {
         for (int j = 0; j < 2 * CHUNK_RENDER_DISTANCE; j++) {
             int x = (int)(cam.position[0] / CHUNK_SIZE) - CHUNK_RENDER_DISTANCE + i;
             int z = (int)(cam.position[2] / CHUNK_SIZE) - CHUNK_RENDER_DISTANCE + j;
-            packets[i][j] = get_chunk_packet(x, z);
+            packets[i][j] = get_chunk_mesh(x, z);
             render_sides(packets[i][j]->opaque_side_data, packets[i][j]->num_opaque_sides);
         }
     }
